♻️ refactor: Move Validations class to helpers folder
